+++ cargo-marker/src/backend/driver.rs
@@ -1,3 +1,4 @@
+use super::cargo::Cargo;
 use super::toolchain::{get_toolchain_folder, rustup_which, Toolchain};
 use crate::error::prelude::*;
 use crate::observability::display::print_stage;
@@ -16,11 +17,13 @@ pub fn marker_driver_bin_name() -> String {
 /// to install the driver.
 pub(crate) fn default_driver_info() -> DriverVersionInfo {
     DriverVersionInfo {
-        toolchain: "nightly-2023-08-24".to_string(),
-        // region replace-version dev
-        version: "0.4.0-dev".to_string(),
-        api_version: "0.4.0-dev".to_string(),
-        // endregion replace-version dev
+        // region replace rust toolchain dev
+        toolchain: "nightly-2023-12-28".to_string(),
+        // endregion replace rust toolchain dev
+        // region replace marker version dev
+        version: "0.6.0-dev".to_string(),
+        api_version: "0.6.0-dev".to_string(),
+        // endregion replace marker version dev
     }
 }
 
@@ -138,11 +141,10 @@ fn build_driver(toolchain: &str, version: &str, mut additional_rustc_flags: Opti
     }
 
     // Build driver
-    let mut cmd = Command::new("cargo");
+    let mut cmd = Cargo::with_toolchain(toolchain).command();
     if is_local_driver() {
         cmd.args(["build", "--bin", "marker_rustc_driver"]);
     } else {
-        cmd.env("RUSTUP_TOOLCHAIN", toolchain);
         cmd.args(["install", "marker_rustc_driver", "--version", version, "--force"]);
 
         *additional_rustc_flags.get_or_insert_with(Default::default) += " --cap-lints=allow";