+++ marker_adapter/src/loader.rs
@@ -1,5 +1,6 @@
 use crate::error::prelude::*;
 use camino::Utf8PathBuf;
+use itertools::Itertools;
 use libloading::Library;
 use marker_api::{LintCrateBindings, MarkerContext};
 use marker_api::{LintPass, LintPassInfo, MARKER_API_VERSION};
@@ -62,6 +63,22 @@ impl LintCrateRegistry {
             new_self.passes.push(LoadedLintCrate::try_from_info(krate.clone())?);
         }
 
+        let lint_passes = new_self.collect_lint_pass_info();
+
+        let errors = lint_passes
+            .iter()
+            .flat_map(LintPassInfo::lints)
+            .into_group_map_by(|lint| lint.name.to_ascii_lowercase())
+            .into_iter()
+            .filter(|(_, lints)| lints.len() > 1)
+            .map(|(lint_name, lints)| {
+                let defs = lints.iter().map(|lint| format!("- {}", lint.fqn)).format("\n");
+
+                Error::root(format!("The lint `{lint_name}` is defined multiple times:\n{defs}",))
+            });
+
+        Error::try_many(errors, "Found several lint name conflicts")?;
+
         Ok(new_self)
     }
 
@@ -80,6 +97,12 @@ impl LintCrateRegistry {
 impl LintPass for LintCrateRegistry {
     fn info(&self) -> LintPassInfo {
         panic!("`registered_lints` should not be called on `LintCrateRegistry`");
+    }
+
+    fn check_crate<'ast>(&mut self, cx: &'ast MarkerContext<'ast>, krate: &'ast marker_api::ast::Crate<'ast>) {
+        for lp in &self.passes {
+            (lp.bindings.check_crate)(cx, krate);
+        }
     }
 
     fn check_item<'ast>(&mut self, cx: &'ast MarkerContext<'ast>, item: marker_api::ast::ItemKind<'ast>) {