+++ marker_uilints/tests/ui/expr/sugar/for_loop.rs
@@ -1,11 +1,8 @@
-#![feature(register_tool)]
-#![register_tool(marker)]
-
 fn main() {
     let range = 1..10;
     let mut total = 0;
 
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     for i in range {
         total += i;
     }