+++ marker_uilints/tests/ui/expr/sugar/ranges.rs
@@ -1,13 +1,10 @@
-#![feature(register_tool)]
-#![register_tool(marker)]
-
 fn main() {
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     let _ = 1..2;
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     let _ = ..2;
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     let _ = 1..;
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     let _ = ..;
 }