+++ marker_uilints/tests/ui/expr/sugar/while_loop.rs
@@ -1,10 +1,7 @@
-#![feature(register_tool)]
-#![register_tool(marker)]
-
 fn main() {
     let mut something = Some(12);
 
-    #[warn(marker::print_every_expr)]
+    #[warn(marker::marker_uilints::print_every_expr)]
     while let Some(_) = something {
         something = None;
     }