+++ scripts/release/qemu-setup.sh
@@ -17,6 +17,8 @@ apt-get-install \
     curl
 
 function curl_with_retry {
+    # FIXME: replace `--retry-connrefused` with `--retry-all-errors` when we
+    # remove support for ubuntu20.04 which has an old curl version.
     with_log curl \
         --location \
         --silent \